etc: move const folders from core folder to util foler
