+++ src/@components/CategoryPage/CategoryContents/index.tsx
@@ -3,8 +3,8 @@ import { useSetRecoilState } from "recoil";
 
 import { useCategoryLists } from "../../../core/api/main";
 import { sliderIdxState } from "../../../core/atom/slider";
-import { gridValue } from "../../../core/category/categoryList";
 import { routePaths } from "../../../core/routes/path";
+import { gridValue } from "../../../util/category/categoryList";
 import { GTM_CLASS_NAME } from "../../../util/const/gtm";
 import Loading from "../../@common/Loading";
 import { St } from "./style";
