+++ src/@components/JoinPage/EmailAuthenticationPage/index.tsx
@@ -10,12 +10,12 @@
 import axios from "axios";
 import { useEffect, useState } from "react";
 
-import { subHeaderInfo } from "../../../core/join/subHeaderInfo";
-// import { useNavigate } from "react-router-dom";
-// import { joinApi } from "../../../core/api/join";
-import { EmailInvalidMessage, emailInvalidMessage } from "../../../core/join/userInfoInputErrorMessage";
 // import { routePaths } from "../../../core/routes/path";
 import checkEmailInvalid from "../../../util/checkInvalidEmail";
+import { subHeaderInfo } from "../../../util/join/subHeaderInfo";
+// import { useNavigate } from "react-router-dom";
+// import { joinApi } from "../../../core/api/join";
+import { EmailInvalidMessage, emailInvalidMessage } from "../../../util/join/userInfoInputErrorMessage";
 import Footer from "../../@common/Footer";
 import { useDebounce } from "../../@common/hooks/useDebounce";
 import SubHeader from "../../@common/SubHeader";
