+++ src/@components/MainPage/BestPiickle/index.tsx
@@ -1,5 +1,5 @@
 import { useBestPiickle } from "../../../core/api/main";
-import { headingTitles } from "../../../core/main/headingTitles";
+import { headingTitles } from "../../../util/main/headingTitles";
 import HeadingTitleContainer from "../../@common/HeadingTitleContainer";
 import Loading from "../../@common/Loading";
 import BestPiickleCard from "./BestPiickleCard";
