+++ src/util/delete/feedBackListsContents.ts
@@ -1,0 +1,7 @@
+export const feedBackListsContents = [
+  { text: "마음에 드는 주제가 없어요" },
+  { text: "이용이 불편해요" },
+  { text: "다른 서비스나 제품을 더 자주 이용해요" },
+  { text: "잘 사용하지 않아요" },
+  { text: "기타" },
+];
