+++ src/util/vote/voteContent.ts
